Moved files around

+++ Hazel/src/Hazel/Renderer/Resources/Buffer.h
@@ -36,7 +36,7 @@ namespace Hazel {
 	{
 		std::string Name;
 		ShaderDataType Type;
-		uint32_t Offset;
+		uint64_t Offset;
 		uint32_t Size;
 		bool Normalized;
